add: new file for CoordinationSemaphore
