+++ coordination/src/main/java/tech/ydb/coordination/impl/CoordinationSessionNewImpl.java
@@ -5,11 +5,8 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.function.Consumer;
 
 import tech.ydb.coordination.CoordinationSessionNew;
-import tech.ydb.coordination.settings.DescribeSemaphoreChanged;
-import tech.ydb.coordination.settings.SemaphoreDescription;
 import tech.ydb.core.Result;
 import tech.ydb.core.Status;
 import tech.ydb.core.StatusCode;
@@ -52,7 +49,8 @@ public class CoordinationSessionNewImpl implements CoordinationSessionNew {
         final int semaphoreId = lastId.getAndIncrement();
         return stream.sendCreateSemaphore(semaphoreName, limit, data, semaphoreId)
                 .thenApply(status -> (status.isSuccess() || status.getCode() == StatusCode.ALREADY_EXISTS) ?
-                        Result.success(new CoordinationSemaphoreImpl(semaphoreName)) :
+                        Result.success(new CoordinationSemaphoreImpl(stream, lastId.getAndIncrement(), semaphoreName,
+                                lastId)) :
                         Result.fail(status));
     }
 
@@ -67,7 +65,8 @@ public class CoordinationSessionNewImpl implements CoordinationSessionNew {
         stream.sendAcquireSemaphore(semaphoreName, count, timeout, true, data, semaphoreCreateId);
 
         return acquireFuture.thenApply(status -> status.isSuccess() ?
-                Result.success(new CoordinationEphemeralSemaphoreImpl(semaphoreName, semaphoreCreateId)) :
+                Result.success(new CoordinationEphemeralSemaphoreImpl(stream, lastId.getAndIncrement(), semaphoreName,
+                        lastId)) :
                 Result.fail(status)
         );
     }
@@ -76,89 +75,6 @@ public class CoordinationSessionNewImpl implements CoordinationSessionNew {
     public void close() {
         if (isWorking.compareAndSet(true, false)) {
             stream.stop();
-        }
-    }
-
-    public class CoordinationSemaphoreImpl implements CoordinationSemaphore {
-        private final String name;
-        private final AtomicBoolean isDeleted = new AtomicBoolean(false);
-        private final int semaphoreId;
-
-        public CoordinationSemaphoreImpl(String name) {
-            this.name = name;
-            semaphoreId = lastId.getAndIncrement();
-        }
-
-        @Override
-        public CompletableFuture<Status> update(byte[] data) {
-            final int id = lastId.getAndIncrement();
-            checkDeleted();
-            return stream.sendUpdateSemaphore(name, data, id);
-        }
-
-        @Override
-        public CompletableFuture<Result<Boolean>> acquire(long count, Duration timeout, byte[] data) {
-            if (data == null) {
-                data = BYTE_ARRAY_STUB;
-            }
-            final int id = lastId.getAndIncrement();
-            checkDeleted();
-            return stream.sendAcquireSemaphore(name, count, timeout, false, data, id);
-        }
-
-        @Override
-        public CompletableFuture<Result<Boolean>> release() {
-            final int id = lastId.getAndIncrement();
-            checkDeleted();
-            return stream.sendReleaseSemaphore(name, id);
-        }
-
-        @Override
-        public CompletableFuture<Result<SemaphoreDescription>> describe(DescribeMode describeMode, WatchMode watchMode,
-                                                                Consumer<DescribeSemaphoreChanged> updateWatcher) {
-            checkDeleted();
-            return stream.sendDescribeSemaphore(name,
-                    describeMode.includeOwners(),
-                    describeMode.includeWaiters(),
-                    watchMode.watchData(),
-                    watchMode.watchOwners(),
-                    updateWatcher, semaphoreId);
-        }
-
-        @Override
-        public CompletableFuture<Status> delete(boolean force) {
-            final int id = lastId.getAndIncrement();
-            checkDeleted();
-            final CompletableFuture<Status> deleteFuture = stream.sendDeleteSemaphore(name, force, id);
-            deleteFuture.whenComplete((status, th) -> {
-                if (th == null && status.isSuccess()) {
-                    isDeleted.set(true);
-                }
-            });
-            return deleteFuture;
-        }
-
-        @Override
-        public void close() {
-            release();
-        }
-
-        private void checkDeleted() {
-            if (isDeleted.get()) {
-                throw new IllegalStateException("Semaphore has already deleted");
-            }
-        }
-    }
-
-    public class CoordinationEphemeralSemaphoreImpl extends CoordinationSemaphoreImpl {
-
-        public CoordinationEphemeralSemaphoreImpl(String name, int id) {
-            super(name);
-        }
-
-        @Override
-        public CompletableFuture<Status> delete(boolean force) {
-            throw new UnsupportedOperationException("Ephemeral semaphore couldn't be deleted.");
         }
     }
 }
